fix: check exists directory before remove it (#2353)

+++ src/MediaCollections/Filesystem.php
@@ -217,7 +217,8 @@ class Filesystem
         collect([$mediaDirectory, $conversionsDirectory, $responsiveImagesDirectory])
             ->each(function (string $directory) use ($media) {
                 try {
-                    $this->filesystem->disk($media->conversions_disk)->deleteDirectory($directory);
+                    if ($this->filesystem->disk($media->conversions_disk)->exists($directory))
+                        $this->filesystem->disk($media->conversions_disk)->deleteDirectory($directory);
                 } catch (Exception $exception) {
                     report($exception);
                 }
